Added the ability to not overwrite the first talk behavior of buffer npcs. This is configurable for eveny buffer npc in buffers.xml through the "direct_first_talk" attribute.
True means that the main.html from yanbuffer will directly open when you
click on the npc.

False means that the default server behavior is maintained. That also
means you need to create html files in "DATAPACK_ROOT/data/html/" for
the npcs. For example if the buffer npc is of type L2Npc and has id 100006,
you have to create a "100006.html" in "DATAPACK_ROOT/data/html/default".

The bypass for the default behavior htmls is "YANBufferNpc".
Example: action="bypass -h YANBufferNpc"